requirements.txt
streamlit
pandas
requests
google-api-python-client
google-generativeai
